[MODIFIED] restructure for actions and labels

+++ repositories/label/label_repository.go
@@ -4,7 +4,7 @@ import "github.com/xbklyn/getgoal-app/entities"
 
 type LabelRepository interface {
 	FindAllLabels() ([]entities.Label, error)
-	FindLabelByID(id int) (entities.Label, error)
+	FindLabelByID(id uint64) (entities.Label, error)
 	FindLabelByName(name string) (entities.Label, error)
 	GetSearchLabel() ([]entities.Label, error)
 	Save(label *entities.Label) error
